Updated box to use input system
Fixed a bug that had been bothering me ever since I discovered it on Innovation Day where the bow was unusable only when using a controller because it used the old input system. At first I thought this was a simple fix and was a result of me not stripping out all the old calls to the old systems. Turns out the way I had the attack input exposed made it impossible to achieve the desired behavior. I exposed the proper input and made the appropriate change in the weapon script